Refactor Model Metadata Retrieval Logic
